pyproject.toml
[project]
name = "gramvaani"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.1",
    "sqlalchemy[asyncio]>=2.0.28",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "cachetools>=5.3.3",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "boto3>=1.34.0",
    "streamlit>=1.32.0",
    "pandas>=2.2.0",
    "python-multipart>=0.0.22",
    "aiosqlite>=0.20.0",
    "sounddevice>=0.5.5",
    "numpy>=2.4.2",
    "sarvamai>=0.1.25",
    "pydub>=0.25.1",
    "num2words>=0.5.14",
    "fastrtc[vad]>=0.0.34",
    "livekit-agents[sarvam,silero,turn-detector]~=1.4",
    "livekit-api~=1.0",
    "openai>=2.24.0",
    "livekit-plugins-aws>=1.4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "anyio>=4.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["src/tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.5",
    "anyio>=4.3.0",
]
